Move common generic slots vars and methods into an abstract base class
