+++ Assets/Scripts/UI/Items/UISlot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 654434a724533e74faa6ef54993363b3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
